add a 'next tier' test runner
Historically, the most commonly used test has been the
`improved-sanity-test`.  But we want to be able to run all of the
other, more specific feature tests after a successful execution of the
i-s-t.

This change introduces a modest Python script that will execute
certain 'next tier' tests in succession.  It provides limited support
for the most commonly used options to `ansible-playbook`.  It will
record the output of each test into separate log files and also keep a
file record of any failed tests.